backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
apscheduler
httpx
stripe
